refactor ngpvanapi-client
